First generation of the convergence method.
MODIFIED FILES:
    AMF_tools/omiAmfAK2.m - mainly now set up to return the total WRF column
density, which is needed to calculate the convergence.

    BEHR_Main/BEHR_main.m - large chunk of code added after the initial
calculation of the AMF, it will now scale the boundary layer of the WRF profile
until the WRF and BEHR columns agree to within 10%. Since changing the WRF
profile changes the BEHR profile, this needs to be done iteratively.

    Documentation/BEHR User Guide/BEHR-User-Guide.{tex,pdf} - included the
meaning of the various scaling flags and warnings.